backend/requirements.txt
fastapi[standard]
sqlalchemy
pydantic
google-auth
python-jose[cryptography]
passlib[bcrypt]
